refactor: use fedx code structure
