--- src/components/Unit/index.js
@@ -1,1 +1,0 @@
-export * from './Unit';
